backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
langsmith>=0.1.133


